Disable, or fix C and C++ /W4 level warnings
When compiling with MSVC under the wxWidgets CMake build system, there were
a lot of notes and warnings being generated. This should eliminate all of those,
allowing the library to compile cleanly.

+++ plutovg/source/plutovg-ft-stroker.c
@@ -332,7 +332,7 @@ static PVG_FT_Error ft_stroke_border_lineto(PVG_FT_StrokeBorder border,
         border->points[border->num_points - 1] = *to;
     } else {
         /* don't add zero-length lineto, but always add moveto */
-        if (border->num_points > border->start &&
+        if ((int) border->num_points > border->start &&
             PVG_FT_IS_SMALL(border->points[border->num_points - 1].x - to->x) &&
             PVG_FT_IS_SMALL(border->points[border->num_points - 1].y - to->y))
             return error;
